Refactor sessions-guide Phase 2: extract shared mutation logic and stabilize popstate listener
- Extract hasTimeConflict, scheduleAction, and favoriteAction into new
  services/session-actions.js, eliminating ~150 lines of duplication
  between SessionCard and LiveCard
- Standardize auth guard to check isLoggedIn before isRegistered with
  toast notifications (LiveCard previously only checked isRegistered
  via a modal prompt)
- Stabilize DrawerShell popstate listener with a sessionsRef so it
  registers once instead of re-registering on every 30s poll cycle
