+++ event-libs/v1/blocks/sessions-guide/components/SessionCard.js
@@ -1,20 +1,6 @@
 import { isSessionOnDemand, formatSessionTime, formatShortTime, getNowMs } from '../utils/time.js';
-import { addSession, removeSession, toggleSessionInterest } from '../services/rainfocus.js';
+import { scheduleAction, favoriteAction } from '../services/session-actions.js';
 import { setSessionParam } from '../utils/url.js';
-
-function hasTimeConflict(a, b) {
-  const aStart = Date.parse(a.startTimeUtc);
-  const aEnd = Date.parse(a.endTimeUtc);
-  const bStart = Date.parse(b.startTimeUtc);
-  const bEnd = Date.parse(b.endTimeUtc);
-  return aStart < bEnd && aEnd > bStart;
-}
-
-function findScheduleConflict(incoming, sessions, scheduled) {
-  return sessions.find(
-    (s) => s.id !== incoming.id && scheduled.has(s.id) && hasTimeConflict(s, incoming),
-  ) || null;
-}
 
 export function buildSessionCard(preact, store) {
   const { html } = preact;
@@ -22,16 +8,9 @@ export function buildSessionCard(preact, store) {
 
   return function SessionCard({ session }) {
     const { state, dispatch } = useSessionGuide();
-    const {
-      scheduled, favorited, isLoggedIn, isRegistered, eventConfig,
-    } = state;
+    const { scheduled, favorited, eventConfig } = state;
     const pendingActions = state.pendingActions || new Set();
-    const sessions = state.sessions || [];
-    const {
-      userTz, surface, trackColors, trackIcons,
-      rfApiProfileId, rfApiUrl, showConflictModal,
-      title: eventTitle, registerUrl,
-    } = eventConfig;
+    const { userTz, surface, trackColors, trackIcons } = eventConfig;
 
     const isScheduled = scheduled.has(session.id);
     const isFavorited = favorited.has(session.id);
@@ -57,98 +36,14 @@ export function buildSessionCard(preact, store) {
       isPending ? 'is-pending' : '',
     ].filter(Boolean).join(' ');
 
-    function showAuthToast(action) {
-      if (isLoggedIn !== true) {
-        dispatch({
-          type: 'SHOW_TOAST',
-          message: `Login required to ${action}`,
-          variant: 'informative',
-          ctaLabel: 'Login to Adobe',
-          ctaAction: () => window.adobeIMS?.signIn(),
-          duration: null,
-        });
-        return true;
-      }
-      if (isRegistered !== true) {
-        const eventName = eventTitle ? ` for ${eventTitle}` : '';
-        dispatch({
-          type: 'SHOW_TOAST',
-          message: `Registration${eventName} required to ${action}`,
-          variant: 'informative',
-          ctaLabel: 'Register',
-          ctaHref: registerUrl || '/register',
-          duration: null,
-        });
-        return true;
-      }
-      return false;
-    }
-
     async function handleSchedule(e) {
       e.stopPropagation();
-      if (showAuthToast('add to schedule')) return;
-      if (isPending) return;
-
-      dispatch({ type: 'SET_PENDING', sessionId: session.id, pending: true });
-      try {
-        if (isScheduled) {
-          // TODO: replace null credentials with real rfAuthToken/clientId from auth integration
-          await removeSession(session.rfCode, null, null, rfApiProfileId, rfApiUrl);
-          dispatch({ type: 'SCHEDULE_REMOVE', sessionId: session.id });
-          dispatch({ type: 'SHOW_TOAST', message: 'Removed from schedule', variant: 'default' });
-        } else {
-          if (showConflictModal) {
-            const conflict = findScheduleConflict(session, sessions, scheduled);
-            if (conflict) {
-              dispatch({
-                type: 'SHOW_CONFLICT',
-                conflict: {
-                  existing: conflict,
-                  incoming: session,
-                  onConfirm: async (keep) => {
-                    if (keep.id === session.id) {
-                      await removeSession(conflict.rfCode, null, null, rfApiProfileId, rfApiUrl);
-                      dispatch({ type: 'SCHEDULE_REMOVE', sessionId: conflict.id });
-                      await addSession(session.rfCode, null, null, rfApiProfileId, rfApiUrl);
-                      dispatch({ type: 'SCHEDULE_ADD', sessionId: session.id });
-                    }
-                    dispatch({ type: 'SHOW_TOAST', message: 'Schedule updated', variant: 'success' });
-                  },
-                },
-              });
-              dispatch({ type: 'SET_PENDING', sessionId: session.id, pending: false });
-              return;
-            }
-          }
-          await addSession(session.rfCode, null, null, rfApiProfileId, rfApiUrl);
-          dispatch({ type: 'SCHEDULE_ADD', sessionId: session.id });
-          dispatch({ type: 'SHOW_TOAST', message: 'Added to schedule', variant: 'success' });
-        }
-      } catch (err) {
-        window.lana?.log(`[sessions-guide] schedule action failed: ${err.message}`);
-        dispatch({ type: 'SHOW_TOAST', message: 'Something went wrong. Please try again.', variant: 'error' });
-      } finally {
-        dispatch({ type: 'SET_PENDING', sessionId: session.id, pending: false });
-      }
+      await scheduleAction(session, state, dispatch);
     }
 
     async function handleFavorite(e) {
       e.stopPropagation();
-      if (showAuthToast('add to favorites')) return;
-      if (isPending) return;
-
-      dispatch({ type: 'SET_PENDING', sessionId: session.id, pending: true });
-      try {
-        // TODO: replace null credentials with real rfAuthToken/clientId from auth integration
-        await toggleSessionInterest(session.rfCode, session.id, null, null, rfApiProfileId, rfApiUrl);
-        dispatch({ type: isFavorited ? 'FAVORITE_REMOVE' : 'FAVORITE_ADD', sessionId: session.id });
-        dispatch({ type: 'SHOW_TOAST', message: isFavorited ? 'Removed from favorites' : 'Added to favorites', variant: isFavorited ? 'default' : 'success' });
-      } catch (err) {
-        window.lana?.log(`[sessions-guide] favorite action failed: ${err.message}`);
-        dispatch({ type: 'SHOW_TOAST', message: 'Something went wrong. Please try again.', variant: 'error' });
-      } finally {
-        dispatch({ type: 'SET_PENDING', sessionId: session.id, pending: false });
-      }
+      await favoriteAction(session, state, dispatch);
     }
 
     function handleClick() {
